* Changed heap id to a full int to support handle-type IDs. * Fixed up some copy/assign allocation issues in collector buffer. * Fixed some destruction issues in collector queue.

+++ Collector/Allocation.h
@@ -19,7 +19,7 @@ namespace Alloclave
 		unsigned int Size;
 		unsigned int Alignment;
 		AllocationType Type;
-		unsigned short HeapId;
+		unsigned int HeapId;
 
 		char* Stack;
